requirements.txt
astunparse==1.6.3
async-generator==1.10
atpublic==2.3
attrs==21.2.0
backcall==0.2.0
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
colorlover==0.3.0
commonmark==0.9.1
configobj==5.0.6
cufflinks==0.17.3
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
dictdiffer==0.8.1
diskcache==5.2.1
distributed==2021.7.1
distro==1.5.0
dpath==2.0.1
dulwich==0.20.21
entrypoints==0.3
flatbuffers==1.12
flatten-dict==0.3.0
flufl.lock==3.2
fsspec==0.9.0
ftfy==6.0.1
funcy==1.15
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.14
grandalf==0.6
grpcio==1.34.1
HeapDict==1.0.1
impyute==0.0.8
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jsonpath-ng==1.5.2
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-contrib-core==0.3.3
jupyter-core==4.7.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.36.0
locket==0.2.1
lxml==4.6.3
mailchecker==4.0.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
matrixprofile==1.1.10
mistune==0.8.4
mpmath==1.2.1
msgpack==1.0.2
nanotime==0.5.2
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numba==0.53.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pathlib2==2.3.5
pathspec==0.8.1
patsy==0.5.1
phonenumbers==8.12.22
pickleshare==0.7.5
plotly==5.1.0
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.11.2
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pydot==1.4.2
pygit2==1.5.0
Pygments==2.9.0
pygtrie==2.3.2
pyodbc==4.0.31
pyparsing==2.4.7
pyrsistent==0.18.0
python-benedict==0.24.0
python-dateutil==2.8.1
python-fsutil==0.5.0
python-slugify==5.0.0
pytrends==4.7.3
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.7.6
requests-oauthlib==1.3.0
rich==10.1.0
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
seaborn==0.11.1
Send2Trash==1.7.1
shortuuid==1.0.1
shtab==1.3.6
sklearn==0.0
smmap==4.0.0
sortedcontainers==2.4.0
statsmodels==0.12.2
stumpy==1.9.1
sympy==1.8
tabula-py==2.2.0
tabulate==0.8.9
tblib==1.7.0
tenacity==7.0.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.1.0
tifffile==2021.4.8
toml==0.10.2
toolz==0.11.1
tornado==6.1
traitlets==5.0.5
tsfresh==0.18.0
Unidecode==1.2.0
voluptuous==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.0
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xmltodict==0.12.0
zc.lockfile==2.0
zict==2.0.0
